# Documents/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FastFood API - Ingenito Emiddio",
    "description": "Documentazione endpoint API REST, per il progetto \"FastFood\" del corso di Programmazione Web e Mobile - A.A. 2025/2026",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/meals/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "strCategory",
            "in": "query",
            "type": "string"
          },
          {
            "name": "strArea",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          },
          {
            "name": "prezzoMin",
            "in": "query",
            "type": "string"
          },
          {
            "name": "prezzoMax",
            "in": "query",
            "type": "string"
          },
          {
            "name": "ristorante_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "strMeal": {
                  "example": "any"
                },
                "strCategory": {
                  "example": "any"
                },
                "strArea": {
                  "example": "any"
                },
                "strInstructions": {
                  "example": "any"
                },
                "strMealThumb": {
                  "example": "any"
                },
                "ingredients": {
                  "example": "any"
                },
                "measures": {
                  "example": "any"
                },
                "prezzo": {
                  "example": "any"
                },
                "tempo_preparazione": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/meals/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}